Run one reflection test per property
Blink/WebKit are having trouble with these tests because they're too
big, even after splitting up.  #4078 makes a slight dent, but it looks
like my original intent was that each property should be one test, but I
didn't wind up implementing it.  This cuts the number of tests run by a
factor of something like 80 (e.g., 15,065 to 191 for embedded).

Naturally, it has the disadvantage that if an implementation fails
multiple assertions in a property, there will be no more test coverage
for that property, neither for regressions nor standards conformance.
In some cases a minor bug that's hard for a certain implementation to
fix could indefinitely disable further test coverage for a wide array of
properties.  As such, I'm not sure this is a good idea.  If the problem
that Blink/WebKit have is that the log messages are too verbose, perhaps
they shouldn't print log lines for expected passes/fails, which I
believe is what Mozilla does (speak to jgraham).

If we do want this change, we could also probably recombine the test
files.